+++ python/lsst/daf/butler/__init__.py
@@ -7,7 +7,7 @@ Data Access Butler
 
 from .core import *
 from .registry import Registry  # import the registry subpackage directly for other symbols
-from .butlerConfig import *
-from .deferredDatasetHandle import *
-from .butler import *
+from ._butlerConfig import *
+from ._deferredDatasetHandle import *
+from ._butler import *
 from .version import *
